requirements.txt
# Core
streamlit>=1.45.0
pytesseract>=0.3.10
Pillow>=10.0.0
transformers>=4.40.0
sentencepiece>=0.2.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# OCR / Imagem
opencv-python>=4.8.0

# PDF
pymupdf>=1.24.0

# Detecção de língua
langid>=1.1.6

# Exportação
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.0.0

# Segmentação de frases (opcional)
nltk>=3.8.0

# Testes
pytest>=7.4.0
